Redeploy schema (#335)
This was done to correct problems with some old files having CRLF line endings.  The revision attribute was set to a unique value for all files in the previous commit/merge.  This now sets the revision atrtribute to the correct value.  The deplpoy processor fixes the CFLF to LF issue for all files.